backend/requirements.txt
altgraph==0.17
astroid==2.3.3
colorama==0.4.3
et-xmlfile==1.0.1
future==0.18.2
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.20.2
openpyxl==3.0.7
pandas==1.2.3
pefile==2019.4.18
pyinstaller==5.0.dev0
pyinstaller-hooks-contrib==2021.1
pylint==2.4.4
PySide2==5.15.2
python-dateutil==2.8.1
pytz==2021.1
pywin32-ctypes==0.2.0
shiboken2==5.15.2
six==1.13.0
wrapt==1.11.2
xlrd==2.0.1
Django==3.1.7
djangorestframework==3.12.4
pyMysql==1.0.2
django-cors-headers==3.7.0
djangorestframework-jwt==1.11.0
cryptography==3.4.7
channels==2.4.0
channels-redis==2.4.2
asgiref==3.3.4